Centralize dependencies for temporary system. Ensure all prior packages present.
I was worried that I may have been linking against system libraries on accident
instead of dependencies in i686-tools or i686-cross that I wasn't aware of. This
is a bit of a shotgun hack of a fix, but it should yield a usable Linux build -
and is likely tolerable in the cross-compilation tools and temporary system.

May be a bit of a cleanup effort in the future, though.