Fix hanging gitzebo boot and shutdown, add more verbose debug output.
Without this, gitzebo started, but failed to initialize all processes due
to https-fileserver starting in the foreground.

It also didn't record all PIDs, leading to an incomplete shutdown.
Incomplete shutdowns trigger the failsafe LXC shutdown behavior, which takes
minimally 2 minutes.  This was annoying.

The debug output was insufficient to figure this issue out, so I added some.