Initial commit of initramfs image.
This is needed to boot any system where you don't know which
device will hold the rootfs.  For example, when booting a live CD
or install CD, you don't know whether the CD drive is IDE, SATA,
or SCSI, and you don't know which drive index if many are
installed.  This allows you to just search for the disk by label,
at the expense of boot time.