Make it possible to add keys to the kernel config.
Prior to this change, it was only possible to modify the values
of a key already in use.

This should make it possible to configure stuff like squashfs,
which only has one of ~five keys present in the default kernel
configuration.

Of course - I decided that squashfs wasn't needed yet, so the
only remnants of that thought are this configuration code...