Add a stage2 OS installer.
This makes it possible to boot from the stage2 CD and turn
an arbitrary writable disk into a writable stage2 disk.
This is needed to build stage3, which needs to write its
results somewhere, especially when considering that memory
is not large enough for all stage3 packages on all supported
deployment targets.