Fix kernel configuration functions.
The kernel would throw out some kernel configuration settings
in the earlier revision of this file, including CONFIG_SCSI_HPSA.
Basically, it removed any settings with unmet dependencies. This
commit should fix the HPSA driver, and make the kernel build fail
when this error recurs in the future.

It will recur, because we do not do dependency resolution for the
configuration values we set - so we can actually set mutually
exclusive configuration settings.

This commit at least makes it clear when it explodes, and
documents why.