Add unit tests for copy_diff_files.py.
The stage2 image does not have a Python interpreter.  The Python
interpreter is annoying to cross-compile.  The stage2 image does
have a bash interpreter, though.  Since the stage2 image needs to
be capable of building new packages this is a problem, so...

copy_diff_files.py is about to be rewritten in bash.  These tests
will make it easier to have some vague reassurances that the
replacement will work.