Prevent unit tests from polluting filesystem with temp dirs.
The cleanup.sh exit handlers don't really do a good job of running
when a subshell exits, so they have to be invoked manually.  Maybe
a valid fix exists - but subshells only really come up in unit
tests, so it's probably not worth the effort to fix at the moment.