Prevent .installed_pkgs filelists from containing misleading . prefix.
Earlier, we might output ./etc/some/config/dir/config.conf for some
file at /etc/some/config/dir/config.conf.  Which directory does the "."
refer to?  It was actually "/" which it referred to, but in a chroot
it might be something else entirely.

The prefix is now stripped out, but no leading slash is provided yet
as "$chroot_dir/$path" form is expected when manipulating these lists.