Allow files to be explicitly marked for packaging.
Prior to this change, only directories could be explicitly marked
for packaging.  Allowing files to be marked this way allows the
construction of "mutated" packages - where input is an existing
package, and output is the exact same package with some minor
changes.

This will be used to modify the GCC configuration during the stage
3 builds.