Prevent conflicts in directories packaged both implicitly and explicitly.
Prior to this patch, a directory containing modified files would be created
implicitly by copy_diff_files.sh and would then choke if added to the list
of explicitly packaged directories, as the directory already existed.

After this patch, mkdir's -p argument is specified, which quietly gobbles
up errors due to this issue.