Port infrastructure-upping scripts to Ubuntu.
create_all_containers should now ensure it's bootstrapping the
network if the network is uninitialized.  This will help reduce
the number of commands required to bootstrap a build.

create_network.sh uses more generic utilities to create bridges,
and should work with Ubuntu.  The bridges won't persist, as
persistent configuration forces the script to be distribution
specific.  It IS distribution-specific now, but hopefully can
be refactored later to not be...