Implement dependency rewriting for multidistro compatibility.
With this change, we can take a Gigayak package name and output
one or more package names corresponding to the stage 1 (host OS)
environment.  One example where this may be useful is tar: in
CentOS and Gigayak, tar pulls in gzip and friends as a dependency.
On Ubuntu, they need to be pulled in separately.  By setting up
dependency rewriting for the name "tar", Ubuntu's schema can be
satisfied.

Other places this may apply are "-dev" / "-devel" packages, which
are named differently across distributions.