Switch Ubuntu package conversion to actually using apt-get.
Prior to this, an "extract the files" approach similar to RPM
conversion was used.  Unfortunately, Ubuntu has a strong emphasis
on "alternatives" symlinks, which manifest as post-install
configuration scripts.  This manifested as super-common binary
names leading to file-not-found errors - for instance, the "cc"
alias commonly assumed by gcc was nowhere to be found.

This change allows apt-get postinstall scripts to run, but also
breaks when attempting to install gcc due to cyclic dependencies.
It attempts to break cyclic dependencies (which apt-get does),
but uses a slightly simpler mechanism: it breaks the cycle at a
random point, and gcc and friends are expecting the cycle to be
broken at a point where no postinstall script is run.

Detecting a package with no postinstall script during cycle
resolution might be very fun, as there's three cases:
1. No postinstall scripts are present anywhere.
2. Postinstall scripts everywhere, and cycle must be broken
   at random.
3. Postinstall scripts somewhere, so a package with none
   must be located for cycle breaking.

The build system doesn't compute dependencies until a build is
already running, which seriously complicates things.  Having
static dependencies would break dynamic packages (such as the
generic buildscripts for pip modules in Python).