Escape dependency names to prevent g++ from failing to match.
This is a standard example of an injection flaw due to unescaped
string concatenation: not escaping 'g++' before concatenating with
a regular expression causes it to try to match unexpected strings.
'g++' is not one of the strings it matches, as 'g++' is not one
or more 'g's.

See also: https://reddit.com/r/gggg