Require pkg-config to build gnutls.
Without this, gnutls builds would fail complaining that libnettle
2.7.1 could not be found:

    Libnettle 2.7.1 was not found.

That's a somewhat misleading error message, given that it was
actually pkg-config that was missing in this case.  ./configure
would attempt to use pkg-config to determine if a valid version
of nettle was present, and used its raw return code as the test
result.  If pkg-config was not installed, it returned a failure
as well...