Destroy existing network configurations when bootstrapping.
This prevents manual modifications from seeping into network
configurations - bootstrap.sh should be a clean slate.  For
instance, if the bridge was manually set to the up state but
automation didn't handle this, then bootstrap.sh would not be
able to catch the issue unless it blew away all network configs
first.  This change makes sure we have fewer hard-to-catch
issues.