Allow apt to find replacements for dependencies using nonexistent names.
Some dependencies point to a package name where no package actually
exists - instead, some other package points at the same name with a
"Replaces: <that name>" relationship.

This is silly, and breaks apt package conversion... so this change
adds a hack to deal with the issue by parsing the output of apt-get
install's helpful error message.