Use dependency translation for pip packages, and depend on Python.
Earlier, pip packages assumed Python was present... because on CentOS,
it was!  yum depends on Python, and is present in CentOS by default due
to being the package manager for CentOS.

As such, pip packages were horrifically broken on Ubuntu, as executing
a binary would fail with "/usr/bin/python not found!".  Oops.