Allow cgroups to be suffixed with "-1" or "-2" when reloading DNS.
Apparently, sometimes quickly recreating a container "example-01"
can lead to its new instance's cgroup names being suffixed with a
"-1".  When this happens to DNS, the DNS server does not get
reload signals anymore, causing it to lose track of new containers.

This patch allows for the "-" in the "-1" suffix to be globbed,
causing the DNS breakages to stop.