Prevent GMP builds from failing when cross compiling Linux.
ABI=32 on a 64-bit host caused messages along the lines of
"error: size of array 'test_array' is negative" and "configure:
error: could not find a working compiler, see config.log for
details".

Looking at the CLFS book, it appears ABI=32 does not exist
in the latest development version, and removing it from the
commandline clears the errors.