Use correct architecture when building kernel configuration.
Building an i386 kernel using a kernel configuration meant for
x86_64 doesn't end well - this ensures that the correct architecture
is passed to the build system when generating kernel configuration
files.

Without this, the kernel build aborts when it encounters an unset
option that applies to the target architecture, as its default
behavior is to prompt.  Since these builds are non-interactive,
it cannot prompt, and instead just keels over dead.