Make apt package converter exit if a name matches multiple packages.
For an instructive example of why, try converting "libglib2.0" on
Ubuntu.  For some reason, that name matches 10 packages during
apt-get download, and causes apt-cache dotty to silently expand
them.  Ideal would be to either figure out how to pass an exact
name to the apt tools, or figure out how to detect and match this
expansion.

However, since package conversion only needs to work for a small
subset of the Ubuntu package universe, and all converted packages
will be promptly thrown away, I'm deferring work on figuring out
how to fix this in an ideal fashion.  Making it break politely
works for now, as it helps indicate that you're targeting the
wrong package name in dependency translation.