Select only one dependency from each "a|b|c" dependency clause.
The first revision of apt.deps.sh assumed that it could install all
dependencies requested in a group when a package asked for any one
package from a group of packages.

This worked for a while, and then it didn't.  Now, it emits the
first dependency that "seems like it could work" - the first matching
dependency that can be found by apt-get.