Apply dependency translation when converting packages.
This allows you to do ./install_pkg.sh --pkg_name=<canonical name>
and have the correct package built and installed for you.

For example, ./install_pkg.sh --pkg_name=openssh-clients will
automatically find and install the package 'openssh-client' for you
on Ubuntu now.  lfs.stage2.sh attempts to install this package
in precisely this way - which is why this is needed.