Add debug facilities for stage2: logging and chroot preservation.
Logs were already being written, but their location was not
announced to the console.  This change announces the log location
so that it can easily be looked at.

It also adds --preserve_chroot to be able to inspect the logs
after the task terminates, or to be able to actually drop into
the chroot and run commands manually.