Ensure packages created by apt conversion have base-ubuntu available.
create_crypto.sh was failing with:

    chroot: /bin/bash: No such file or directory

This was caused by libc.so.6 not being found, as it's a part of the
base Ubuntu chroot, which managed to dodge installation.  Since all
apt conversion packages are diffed against base-ubuntu when figuring
out what to package, files present in base-ubuntu do not get packaged
into the converted package.  bash properly depended on the libc6
package, but the file libc.so.6 was not present in it - it was
present in base-ubuntu.  If libc6 was marked as depending on
base-ubuntu, this would not have been an issue.