Deprecate ISO creation in favor of working raw image flow.
Using ISO images on remote hosts with remote access cards is nice,
but it required a lot of hacks.  Instead, using a modern processor
with KVM-compatible hardware virtualization extensions seems like
a much less-hacky approach, and one which will be more widely
supported.

The only reason I didn't go with that earlier was that my last
machine did not have these extensions, so I bought a computer
that was more recently built.