Add architecture and distribution name to dependency schema.
This allows architecture and distribution name to be moved out
of package names - so "qemu" can just be called "qemu",
regardless of whether it was built on an Ubuntu host or a Gigayak
host.  This is the first step to making Gigayak bootstrapped /
self-hosting, as it allows one package specification to serve
multiple architectures or host distributions.  This in turn allows
the cross compilation toolchain to be easily built using packages
built for Gigayak, as Gigayak packages are no longer named stuff
like "i686-yak-gcc" - so depending on "gcc" will work even on
Gigayak now.