Allow buildtools to use buildsystem libraries.
This is done somewhat bluntly by installing the entire buildsystem
into the build chroot, up to and including the libraries.

Not all libraries and scripts are guaranteed to work within the
chroot, in fact, some are expected to fail.  Which do or do not
work is left as an exercise to the programmer.