Add subdirectory support to DIR meme.
This required a rewrite of the whole meme, as in environment
variable form, a source'd script could and would happily overwrite
the variable, which was in global scope.

By definining DIR as a function, it can yield the caller's
directory, which was expected wherever `$DIR` was written.

This change was created with:

```bash
find . -path ./cache -prune -o -type f -iname '*.sh' -print \
  | xargs -I{} sed -r \
    -e 's/\$DIR/$(DIR)/g' \
    -e 's/^DIR[=\(].*$/DIR(){(cd "$(dirname "${BASH_SOURCE[1]}")" \&\& pwd)}/g' \
    -i {}
```