Disambiguate between tarball output and buildsystem/ directories.
This punts on abstracting out the extraction logic for now - but
a major refactor is going to have to happen soon to ensure that
the working directory is in a well-known location.  Duplicating
the working directory wildcards between make and install specs is
getting rather silly as other stuff gets added to /root/ (and
/root/ really isn't the place for any of it, either - but that's
yet another refactor).