Use the local buildsystem for tools-buildsystem.
If running from a checked-in buildsystem, this has no effect.

If running from a local buildsystem with uncommitted changes, this
will allow those changes to propagate into the tools image when
the tools-buildsystem package is rebuilt.

It also allows the tools-buildsystem package to build from the
`bootstrap.sh` script, when Gitzebo does not necessarily have a
buildsystem repository created.  Hopefully, this removes another
dependency on the deprecated Gitzebo package, allowing it to be
replaced with something else in the future.