Prevent set -u from causing buildtools sourcing to fail.
At some point, will have to do a bulk change of all set -Eeo...
lines to set -Eeuo pipefail, and then just clean up everything
that breaks as a result.  Now is not that point, though - so
standardizing on set -Eeo pipefail prevents breakages when
sourcing scripts that don't match that convention.

I wish these options could be set on a file-specific level -
for example "set -u for this source file only".