Provide sget to the yak build environments.
This prepares for fetching dependencies inside the environments
using a captive package manager, rather than having the package
manager act from outside of the environment.

yak-wget acts as an example of how this will work in the future,
as it was necessary to choose "whatever is available".  This will
be common for yak packages, as they should compile using the yak
distribution when available, and the tools2 distribution when not.
This way, yak package specifications can be used for both stage3
(native from tools2) compilation and full native compilation stages.