Prevent internal-ca-certificates from packaging bundle file.
Go will look for /etc/ssl/certs/ca-certificates.crt, which this
package will erroneously include prior to this patch.  Go then
stops looking for further sources of CA data.

This instead tries to package files that mention "gigayak" - which
should include the certificate itself and any symlinks to it, which
may be generated on non-Gigayak bootstrapping OSes such as RHEL or
Ubuntu.