Add yak -> Ubuntu translations for headers/etc.
RHEL used NNNN-devel as the name for packages that contained headers
and the like.  yak will package them with the library itself, so wherever
NNNN-devel pointed at NNNN-dev for Ubuntu, NNNN needs to point at NNNN-dev
in the translations file as well.  This allows all packages to switch to
using yak's dependency names.