Factor out distribution configuration a bit.
This makes it easier to tinker with configuration variables in unit
test contexts, and makes it easier to factor out stuff like the
cluster's domain and resulting derived URLs.

I've held off on it until now due to worries that it would add too
much state.  I've tried to keep scripts fairly stateless - so care
has to be taken to test side effects of configuration variables...