Add a logging library for a future refactor.
All logging code going forward can call this library, so that it's
easy to change the log output format based on configuration or the
like.

This should make it easier to save logs, filter logs, and add more
metadata to the logging format.

In fact, it already adds metadata to the existing logging format:
this library outputs the logging level as well as details about the
process in which the log line was called, to help disambiguate
multiple invocations of the same shell script.