Add KVM to kernel configuration and add logging verbosity.
kconfig_finalize function would output only a single unsupported
flag before this patch.  Now, it scans for all flags that were not
successfully set, outputs them, and then dies.  This makes fixing N
unsupported flags an O(1) process rather than an O(N) process,
which helps a lot when considering that it takes ~5-10 minutes to
get to the point where the error messages are dumped to screen.