Add kmod symlinks, instead of trying to repackage them.
It turns out that the kmod build does not create them at all: I
originally thought that they were simply pre-existing, and that
copy_diff_files was not noticing that they had been modified.

Not the case: make install does not create them for kmod - they
have to be created manually.