Refactor container subnet configuration to config.sh values.
This prevents a conflict if the host is connected to a subnet that
matches the container subnet, as the container subnet can be easily
reconfigured to a different IP.  This comes up when using VMs with
bridged networking - as if the VMs are bound to a network configured
by this buildsystem, and they attempt to configure their own network
using the same configuration, then both the host and guest will be
quibbling over that network address space.

Instead, if testing self-hosting builds in a VM, one should now edit
the .yakrc.sh file within the VM to point to a different IP range.