Add network destruction and localstorage initialization to bootstrap.
These were apparently done manually up until now. Caused some
frustration when I changed the subnet, which broke the network due
to the bridge being bound to one subnet, with containers bound to
another.

DNS server wound up not being configured due to lack of a call to
initialize_localstorage - which was discovered when trying to figure
out why DNS didn't work.  (Hint: it was the network issue above...)