Follow symlinks when deciding if/where to create mount targets.
Without this change, broken symlinks could cause major issues:
if $LOCALSTORAGE/$CONTAINER/$MOUNT was a symlink pointing to a
non-existent directory $NONEXISTENT, then trying to create a
directory at $LOCALSTORAGE/$CONTAINER/$MOUNT with mkdir -pv fails
despite the -p flag, complaining that the directory in question
doesn't exist.  This is because it attempts to follow the broken
symlink, which results in an error.

Instead, manually following the symlink and doing mkdir -pv on
$NONEXISTENT correctly creates the target directory.