Support multiple pkgspec directories.
You can now modify the OS configuration to add additional directories
containing package specifications to PKGSPEC_DIRS.

This allows for a repository to be generated from multiple sources,
including additional upstream sources (though at the moment, you'd still
be responsible for syncing them down with git clone or the like), or local
development (which is the current intent).

This is expected to be most useful by allowing you to author env- packages
of your own to use for containers or workspaces.